Requires stripe node library and does set-up
I npm-installed stripe, following documentation at
`https://stripe.com/docs/checkout/express`.  I saved the
publishable key and the secret key to .env. I set up two routes
for stripe, per the documentation. The index route renders the
Checkout form and displays it to the user.  The charge route
retrieves the email address and card token from the POST request
body.  It uses those parameters to create a Stripe customer.
Next, it invokes the `stripe.charges.create` method, providing
the Customer object as an option. The Stripe Node.js client library
performs those operations asynchronously, which means that you need
to use either callbacks or promises to handle the results and ensure
that the functions are invoked in the correct order.