Revamps 'order' schema and sets up CRUD on orders
We did away with our cart collection. We will build the
orders object on the front-end, and pass that object to our
orders route of our API. Creating the order is POST,
DELETING an order is DESTROY.  A user can rate their orders,
which will be PATCH.  Order history will be index (GET).